+++ test/multi_caching.unit.js
@@ -23,10 +23,11 @@ describe("multiCaching", function() {
     var key;
     var memoryTtl;
     var name;
-    var ttl = 5;
+    var defaultTtl;
 
     beforeEach(function() {
         memoryTtl = 0.1;
+        defaultTtl = 5;
 
         memoryCache = caching({store: 'memory', ttl: memoryTtl});
         memoryCache2 = caching({store: 'memory', ttl: memoryTtl});
@@ -47,7 +48,7 @@ describe("multiCaching", function() {
 
         describe("set()", function() {
             it("lets us set data in all caches", function(done) {
-                multiCache.set(key, value, ttl, function(err) {
+                multiCache.set(key, value, {ttl: defaultTtl}, function(err) {
                     checkErr(err);
 
                     memoryCache.get(key, function(err, result) {
@@ -69,7 +70,7 @@ describe("multiCaching", function() {
             });
 
             it("lets us set data without a callback", function(done) {
-                multiCache.set(key, value, ttl);
+                multiCache.set(key, value, {ttl: defaultTtl});
                 setTimeout(function() {
                     multiCache.get(key, function(err, result) {
                         checkErr(err);
@@ -94,7 +95,34 @@ describe("multiCaching", function() {
                 }, 20);
             });
 
-            it("lets us set data without a ttl or callback", function(done) {
+            it("lets us set data without an options param", function(done) {
+                multiCache.set(key, value, function(err) {
+                    checkErr(err);
+
+                    multiCache.get(key, function(err, result) {
+                        checkErr(err);
+                        assert.equal(result, value);
+
+                        memoryCache.get(key, function(err, result) {
+                            checkErr(err);
+                            assert.equal(result, value);
+
+                            memoryCache2.get(key, function(err, result) {
+                                checkErr(err);
+                                assert.equal(result, value);
+
+                                memoryCache3.get(key, function(err, result) {
+                                    checkErr(err);
+                                    assert.equal(result, value);
+                                    done();
+                                });
+                            });
+                        });
+                    });
+                });
+            });
+
+            it("lets us set data without options or callback", function(done) {
                 multiCache.set(key, value);
                 setTimeout(function() {
                     multiCache.get(key, function(err, result) {
@@ -123,7 +151,7 @@ describe("multiCaching", function() {
 
         describe("get()", function() {
             it("gets data from first cache that has it", function(done) {
-                memoryCache3.set(key, value, ttl, function(err) {
+                memoryCache3.set(key, value, function(err) {
                     checkErr(err);
 
                     multiCache.get(key, function(err, result) {
@@ -133,11 +161,31 @@ describe("multiCaching", function() {
                     });
                 });
             });
+
+            it("passes any options to underlying caches", function(done) {
+                multiCache.set(key, value, function(err) {
+                    checkErr(err);
+
+                    sinon.spy(memoryCache.store, 'get');
+
+                    var opts = {foo: 'bar'};
+
+                    multiCache.get(key, opts, function(err, result) {
+                        checkErr(err);
+
+                        assert.equal(result, value);
+                        assert.ok(memoryCache.store.get.calledWith(key, opts));
+
+                        memoryCache.store.get.restore();
+                        done();
+                    });
+                });
+            });
         });
 
         describe("del()", function() {
             it("lets us delete data in all caches", function(done) {
-                multiCache.set(key, value, ttl, function(err) {
+                multiCache.set(key, value, function(err) {
                     checkErr(err);
 
                     multiCache.del(key, function(err) {
@@ -162,7 +210,7 @@ describe("multiCaching", function() {
             });
 
             it("lets us delete data without a callback", function(done) {
-                multiCache.set(key, value, ttl, function(err) {
+                multiCache.set(key, value, function(err) {
                     checkErr(err);
 
                     multiCache.del(key);
@@ -201,7 +249,7 @@ describe("multiCaching", function() {
             });
 
             it("gets data from first cache that has it", function(done) {
-                memoryCache3.set(key, value, ttl, function(err) {
+                memoryCache3.set(key, value, function(err) {
                     checkErr(err);
 
                     multiCache.getAndPassUp(key, function(err, result) {
@@ -257,7 +305,7 @@ describe("multiCaching", function() {
             });
 
             it("checks to see if higher levels have item", function(done) {
-                memoryCache3.set(key, value, ttl, function(err) {
+                memoryCache3.set(key, value, function(err) {
                     checkErr(err);
 
                     multiCache.getAndPassUp(key, function(err, result) {
@@ -280,9 +328,9 @@ describe("multiCaching", function() {
                 var memoryStoreStub;
 
                 beforeEach(function() {
-                    memoryStoreStub = memoryStore.create({ttl: ttl});
+                    memoryStoreStub = memoryStore.create({ttl: defaultTtl});
                     sinon.stub(memoryStore, 'create').returns(memoryStoreStub);
-                    memoryCache = caching({store: 'memory', ttl: ttl});
+                    memoryCache = caching({store: 'memory', ttl: defaultTtl});
                     multiCache = multiCaching([memoryCache]);
                     fakeError = new Error(support.random.string());
                     sinon.stub(memoryStoreStub, 'get').yields(fakeError);
@@ -318,13 +366,17 @@ describe("multiCaching", function() {
                     memoryCache3.store.set.restore();
                 });
 
+                /**
+                 * Note: it's up to the underlying cache implementation to handle the ttl number.
+                 * We're just testing that the ttl gets passed to the underlying store.
+                 */
                 it('when a ttl number is passed in', function(done) {
                     multiCache.wrap(key, function(cb) {
                         methods.getWidget(name, cb);
-                    }, ttl, function(err, widget) {
+                    }, defaultTtl, function(err, widget) {
                         checkErr(err);
                         assert.deepEqual(widget, {name: name});
-                        sinon.assert.calledWith(memoryCache3.store.set, key, {name: name}, ttl);
+                        sinon.assert.calledWith(memoryCache3.store.set, key, {name: name}, defaultTtl);
                         done();
                     });
                 });
@@ -332,15 +384,15 @@ describe("multiCaching", function() {
                 it('when a ttl option is passed in', function(done) {
                     multiCache.wrap(key, function(cb) {
                         methods.getWidget(name, cb);
-                    }, {ttl: ttl}, function(err, widget) {
+                    }, {ttl: defaultTtl}, function(err, widget) {
                         checkErr(err);
                         assert.deepEqual(widget, {name: name});
-                        sinon.assert.calledWith(memoryCache3.store.set, key, {name: name}, {ttl: ttl});
+                        sinon.assert.calledWith(memoryCache3.store.set, key, {name: name}, {ttl: defaultTtl});
                         done();
                     });
                 });
 
-                it('when a ttl is not passed in', function(done) {
+                it('when no options are passed in', function(done) {
                     multiCache.wrap(key, function(cb) {
                         methods.getWidget(name, cb);
                     }, function(err, widget) {
@@ -381,7 +433,7 @@ describe("multiCaching", function() {
                     function getCachedFalseyValue(cb) {
                         multiCache.wrap(key, function(cacheCb) {
                             getFalseyValue(cacheCb);
-                        }, ttl, cb);
+                        }, cb);
                     }
 
                     beforeEach(function(done) {
@@ -442,7 +494,7 @@ describe("multiCaching", function() {
                         getValue(name, function(err, result) {
                             cacheCb(err, result);
                         });
-                    }, ttl, cb);
+                    }, cb);
                 }
 
                 beforeEach(function() {
@@ -521,7 +573,7 @@ describe("multiCaching", function() {
 
             context("when value exists in first store but not second", function() {
                 it("returns value from first store, does not set it in second", function(done) {
-                    memoryCache.set(key, {name: name}, ttl, function(err) {
+                    memoryCache.set(key, {name: name}, function(err) {
                         checkErr(err);
 
                         multiCache.wrap(key, function(cb) {
@@ -542,7 +594,7 @@ describe("multiCaching", function() {
 
             context("when value exists in second store but not first", function() {
                 it("returns value from second store, sets it in first store", function(done) {
-                    memoryCache3.set(key, {name: name}, ttl, function(err) {
+                    memoryCache3.set(key, {name: name}, function(err) {
                         checkErr(err);
 
                         multiCache.wrap(key, function(cb) {
@@ -604,7 +656,7 @@ describe("multiCaching", function() {
 
             context("when value exists in first store only", function() {
                 it("returns value from first store, does not set it in second or third", function(done) {
-                    memoryCache.set(key, {name: name}, ttl, function(err) {
+                    memoryCache.set(key, {name: name}, function(err) {
                         checkErr(err);
 
                         multiCache.wrap(key, function(cb) {
@@ -630,7 +682,7 @@ describe("multiCaching", function() {
 
             context("when value exists in second store only", function() {
                 it("returns value from second store, sets it in first store, does not set third store", function(done) {
-                    memoryCache3.set(key, {name: name}, ttl, function(err) {
+                    memoryCache3.set(key, {name: name}, function(err) {
                         checkErr(err);
 
                         multiCache.wrap(key, function(cb) {
@@ -656,7 +708,7 @@ describe("multiCaching", function() {
 
             context("when value exists in third store only", function() {
                 it("returns value from third store, sets it in first and second stores", function(done) {
-                    memoryCache2.set(key, {name: name}, ttl, function(err) {
+                    memoryCache2.set(key, {name: name}, function(err) {
                         checkErr(err);
 
                         multiCache.wrap(key, function(cb) {