+++ test/support.js
@@ -90,9 +90,8 @@ var support = {
     testSetGetDel: function(cache, cb) {
         var key = 'TEST' + support.random.string();
         var val = support.random.string();
-        var ttl;
 
-        cache.set(key, val, ttl, function(err) {
+        cache.set(key, val, function(err) {
             if (err) { return cb(err); }
 
             cache.get(key, function(err, result) {